Napravljene promene na integracionim testovima
