+++ patients/src/test/java/raf/bolnica1/patient/cucumber/medicalexaminationservice/MedicalExaminationServiceTests.java
@@ -1,4 +1,4 @@
-package cucumber.patientservice;
+package raf.bolnica1.patient.cucumber.medicalexaminationservice;
 
 import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
@@ -9,7 +9,7 @@ import static io.cucumber.core.options.Constants.GLUE_PROPERTY_NAME;
 
 @Suite
 @IncludeEngines("cucumber")
-@SelectClasspathResource("features/patientservice")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "raf.bolnica1.patient.services.cucumber.patientservice")
-public class PatientServiceTests {
+@SelectClasspathResource("features/medicalexaminationservice")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "raf.bolnica1.patient.cucumber.medicalexaminationservice")
+public class MedicalExaminationServiceTests {
 }
