+++ patients/src/test/java/raf/bolnica1/patient/cucumber/medicalexaminationservice/MedicalExaminationServiceTestsConfig.java
@@ -1,9 +1,9 @@
-package cucumber.patientservice;
+package raf.bolnica1.patient.cucumber.medicalexaminationservice;
 
 import io.cucumber.spring.CucumberContextConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @CucumberContextConfiguration
 @SpringBootTest
-public class PatientServiceTestsConfig {
+public class MedicalExaminationServiceTestsConfig {
 }
