+++ patients/src/test/java/raf/bolnica1/patient/cucumber/patientcrudservice/PatientCrudServiceTests.java
@@ -1,4 +1,4 @@
-package cucumber.patientcrudservice;
+package raf.bolnica1.patient.cucumber.patientcrudservice;
 
 import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
@@ -10,6 +10,6 @@ import static io.cucumber.core.options.Constants.GLUE_PROPERTY_NAME;
 @Suite
 @IncludeEngines("cucumber")
 @SelectClasspathResource("features/patientcrudservice")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "raf.bolnica1.patient.services.cucumber.patientcrudservice")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "raf.bolnica1.patient.cucumber.patientcrudservice")
 public class PatientCrudServiceTests {
 }
